no need for dedicated primitive mmio ptr

+++ src/build.sh
@@ -408,7 +408,7 @@ i686-elf-g++ -T linker.ld -o "$BUILD_DIR/myos.elf" \
 
 printf "\n\n====== End of Linking =====\n\n"
 
-objdump -D -h "$BUILD_DIR/myos.elf" >"$BUILD_DIR/myos.dump"
+objdump -D -h -M intel "$BUILD_DIR/myos.elf" >"$BUILD_DIR/myos.dump"
 
 objdump -D -h -M intel "$BUILD_DIR/apic.o" >"$BUILD_DIR/apic.dump"
 objdump -D -h -M intel "$BUILD_DIR/special_pointers.o" >"$BUILD_DIR/special_pointers.dump"
@@ -550,6 +550,8 @@ else
 	wait $VNC_PID
 
 	# After you close the VNC viewer, kill QEMU
-	kill $QEMU_PID 2>/dev/null
+	if [[ "$DEBUG_OR_RELEASE" == "release" ]]; then
+		kill $QEMU_PID 2>/dev/null
+	fi
 fi
 # qemu-system-i386 -kernel ./build/myos.elf & # or do this to use the binary directly # -cdrom "$BUILD_DIR/myos.iso" # -kernel "$BUILD_DIR/myos.elf" \
